+++ sandbox/index.ts
@@ -51,7 +51,7 @@ const DEFAULT_SERVER_CONFIG: SandboxNetworkOptions = {
   gasLimit: 6721975,
   defaultBalanceEther: 1000,
   totalAccounts: 10,
-  hardfork: 'cancun',
+  hardfork: 'prague',
   hdPath: 'm/44\'/60\'/0\'/0/',
   mnemonic: 'mimic dune forward party defy island absorb insane deputy obvious brother immense',
   chainId: 1337,
@@ -104,8 +104,10 @@ export class Sandbox {
     if (options.verbose) {
       debug.enable('UNS:sandbox*');
       anvilOptions.silent = false;
+      anvilOptions.verbose = true;
     } else {
       anvilOptions.silent = true;
+      anvilOptions.verbose = false;
     }
 
     if (options.rebuild) {